requirements.txt
# Base
Flask==3.0.0
flask-restful==0.3.10
flask-cors==4.0.0

# Base de données
SQLAlchemy==2.0.25

# Tâches asynchrones
celery==5.3.6
redis==5.0.1

# Utilitaires
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
beautifulsoup4==4.12.2

# Réseaux sociaux
google-api-python-client==2.108.0
facebook-sdk==3.1.0

# FastAPI
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.26.2
python-multipart==0.0.6
pydantic==2.5.2
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Analyse des données et traitement du texte
pandas==2.1.4
textblob==0.17.1
emoji==2.9.0
scikit-learn==1.3.2
